Remove pointer manual test from the build
Followup to 8503f884bbdb50c4bebc8f8a9fce05275b0612b1: much of the qml
was moved to an example, and the leftovers in this directory no longer
include the main.cpp/main.qml/CMakeLists.txt for building a standalone
executable. It's generally enough to run them individually with the qml
runtime, on desktop platforms at least.

Pick-to: 6.3
Task-number: QTBUG-101451
Change-Id: I4b329d0fabb398308c411246f34789d66e7eda33
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>