Fix sub-menu shortcuts not being triggered
When QQuickShortcutContext::matcher is checking who should get a
shortcut, it starts with the object that registered the shortcut and
climbs up its parent hierarchy until it finds a window. Sub-menus,
unlike top-level menus, will not have an associated window until their
parent menu is opened. So, check if the popup is a sub-menu so that
actions within it can grab shortcuts.

Change-Id: I07172b1038af1e9fffd9983f7bf0a155b11bb1c3
Fixes: QTBUG-101034
Pick-to: 6.2 6.3
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>