qqmlimport: Log when no qmldir file is found
Also log when it is found, with the path where it is found.

As a drive-by, add a 'version' label before printing the version of an
imported library. In a static build where some libraries end up with a
version set to 'invalid', it's clearer that the version is invalid,
and not the library.

Pick-to: 6.2 6.3
Change-Id: I998776b13bfe98f1668790419f1102e081878c99
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>