FileDialog: do not add default suffix when content scheme is used
On Android when a file is selected using the default file manager,
the selected file is reported using the url with "content" scheme. This
url does not use file suffixes.
Before this patch, if the FileDialog had a defaultSuffix specified, that
suffix was added to the returned content url, and as a result the file
could not be found.

This problem exists FileDialog from both QtQuick.Dialogs and
Qt.labs.platforms, so this patch implements the similar fix in both of
them.

Fixes: QTBUG-94391
Pick-to: 6.3 6.2
Change-Id: Ic6ba95e301857d9b72ee6f5ddb819b9aae9e66e3
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>