iOS Style: Add QQuickIOSTheme
Add QQuickIOSTheme to initialize the QQuickTheme and set a palette with
the appropriate colors for the iOS style. Use the UIKit's API to get
correct colors in iOS, but hardcode them for macOS, just so we can keep
compiling the iOS style on macOS too.

Note: Ideally, we would use the QPlatformTheme (QIOSTheme) to do this,
but we use a fusion style palette in there because we use the Fusion
style on iOS for widget applications. In the future, we could revisit
this and set a "native-looking" palette to the QIOSTheme, like we are
doing now for the QQuickIOSTheme.

Change-Id: I661ffe5ac3d57b750994f7b46abb9502fd311f44
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>