QQuickTableView: implement PositionMode, deprecate usage of Qt::Alignment
As we need to support more enum options for the second
argument to positionViewAtCell() (like "Visible" and
"Contain"), using Qt::Alignment alone is no longer enough.

This patch will therefore follow the same API as in ListView and
GridView, and switch to use PositionMode. The enum values
will however differ from ListView::PositionMode, since values like
"beginning" and "end" are not applicable to a two dimensional
table view.

To support backwards compatibility when used from QML, the
values will match the values in Qt::Alignment. So e.g
Qt.AlignLeft can be used instead of TableView.AlignLeft.
Since QQuickTableView is a private class, the same concerns are
not relevant for the c++ API.

[ChangeLog][QtQuick][TableView] The function positionViewAtCell()
now takes TableView.PositionMode instead of Qt.Alignment as the
second argument. For backwards compatibility, Qt.Alignment can
still be used.

Change-Id: Ib99262035ef43186d911c2fc2e0420205adbbf84
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>