QQuickTableView: refactor scrollToCell() into two separate functions
ScrollToCell() is a large function, and when used, we always
scroll either horizontal or vertical. Whenever we need to
scroll in both directions, we end up calling the function
twice instead. So to simplify things (especially as a
preparation for later patches that implements support for
more PositionModes), split it up into two functions.

Change-Id: Ie4b5085d0ed9f782d4d481a8de56ea316161409c
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>