QQuickTableView: refactor positionViewAtCell() into two functions
Rather than having one positionViewAtCell() (not the public
API, but the private helper function), split it up
into two functions instead, one for row and one for column.
From all the places we call this function, we already
know which direction we want to position the view
anyway. This is also a preparation to ease the implementation
for the upcoming support for new PositionModes.

Change-Id: I2f5593cd85c7cd165cbec6fb999867380b8fe598
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>