QQuickTableView: for cells bigger than the view, prefer showing the top-left part
When positioning on top a cell, PositionMode::Contain promises to make
the entire cell visible. However, this is not possible for cells that
are larger than the view.

This patch will implement support for being able to specify that
positionViewOnCell should prefer to align the cell bottom and/or
right inside the view, but fall back to top and/or left if the cell
cannot fit inside it. This is useful when navigating the current
index around with the keyboard.

Change-Id: I1d1f053dacee1ffd6c657254130066ab8fe9ecb6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>