iOS Style: Set appropriate color if the button doesn't draw background
The UIButton can have different configurations (styles), one of which
being a filled button style, which draws a background on the button,
and a plain style that doesn't, making the button look more like a label.

For our Button control in the iOS style, we opted to go with the latter
look (plain style, no background) for the flat button, and draw a
background for the normal state. In that case, draw the button text
in the appropriate colors, similar to the native look.

Change-Id: Ic19c2eb82a3184869af26fcf2980d94d6ceeb189
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>