QQuickTableView: change order of row and column
The modelIndex() function is added for Qt 6.4, and
therefore not yet in any release.
Looking at the other functions in QQuickTableView that
takes a row and column as arguments (e.g
positionViewAtCell() and itemAtCell()), they all have
column as first argument, and row as second. So change
the modelIndex() function to do the same, so that the
TableView API ends up consistent.

Change-Id: Ia85ceb3a4719c1e3c16e2023aa039263fef83f6e
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>