QQuickTableView: add new function: cellAtPosition()
The current "cellAtPos(pos)" function expects pos to be
relative to TableView. The problem is that if you
attach a Pointer Handler to TableView, it will actually
install itself on the contentItem, and the positions it
reports will be relative to the contentItem.

Since the position used in a call to cellAtPos() is
almost always coming from a Pointer Handler attached to
TableView, it makes more sense to replace the cellAtPos()
with a version that expects pos to be relative to the
contentItem instead. This will remove a source of
confusion and discrepancy, and also avoid the need to
convert the pos to and from the contentItem all the
time (also in QQuickTableView's own code).

This patch will therefore add a new function "cellAtPosition()",
and at the same time, deprecate the old "cellAtPos()".

[ChangeLog][QtQuick][TableView] cellAtPos(pos) is now
deprecated in favor of cellAtPosition(pos). The latter will
assume pos to be relative to the contentItem.

Change-Id: Iaf2f3ae81592bec4370e620c990045fdec6213b8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>