QQmlJSScope: Rework type resolution checks
Extend QQmlJSScope::isFullyResolved() with an extra check for non
composite base. Rearrange existing conditions so that it easier to see
where we consider the type to be resolved

Go over the QQmlJSImportVisitor logic that uses isFullyResolved() and
make sure it behaves properly. Introduce a new function that guarantees
that we report an unresolved type when we see one (where necessary) and
that this is done only once

Change-Id: Ie3c0e9da70fd9e6b2da0075aade8d1d4730bf704
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>