QQuickListView: relayout when section geometry changes
Section delegate items will now notify the ListView when their geometry
changes. Enabling the ListView to call layoutVisibleItems to rearrange
the delegate items based on the geometric changes made to the section
delegate.

The listener is added when the item is created or fetched from the
cache.

The listener is removed when the item is hidden or removed from the
scene.

Fixes: QTBUG-101985
Fixes: QTBUG-100002
Pick-to: 6.2 6.3 6.4 6.5
Change-Id: Ia7617424a8447ba82f62629d93b6a000a72bfc02
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>