QQuickListView: relayout when section geometry changes
Section delegate items will now notify the ListView when their geometry
changes. Enabling the ListView to call layoutVisibleItems to rearrange
the delegate items based on the geometric changes made to the section
delegate.

The listener is added when the item is created or fetched from the
cache.

The listener is removed when the item is hidden or removed from the
scene.

Fixes: QTBUG-101985
Fixes: QTBUG-100002
Change-Id: Ia7617424a8447ba82f62629d93b6a000a72bfc02
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit b994c7ffaf12373ac568a38e3968d620a07c3052)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>