QQmlImport: Clarify addImportToNamespace
The code checker believes the import could leak there. Indeed there was
a hypothetical code path through the function that would leak: If all
existing imports were implicit and the new one was of low precedence.
Rephrase the code to account for this case and make it easier to see
that no code path returns the import without inserting into the list.

Change-Id: I4cd98633fd28d8f5fbc881851eda92f02c9cd1f9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>