FolderDialog: right-align the button box for Basic and Universal styles
The button box in the footer wouldn't be right-aligned by default, for
these styles.

This might be a bit of an aesthetic change, but I think it looks better
when they are right-aligned.

Furthermore, it "fixes" an issue where the cancel button would be
significantly larger than intended for these two styles.
(I quoted the word fixes, because I feel that this is just fixing a
symptom, and not the root cause)

Change-Id: Ie87cc7abad741ddb18187f32212b10d99af2c201
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 6fa855cc75a8d5c569951053c676d23559e77f2e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>