rendercontrol: Correctly adopt a changed external QRhi upon reinit
If we created and own the QRhi, it is then important to not create a new
one in an initialize() that follows an invalidate().

However, if the QRhi was provided from the outside, is can very well
happen that we get a sequence like
initialize-invalidate-now_want_to_use_a_different_rhi-initialize
in which case skipping the rhiSupport->createRhi (which in this case
does not actually create a QRhi but rather wraps an existing one)
is just wrong.

Task-number: QTBUG-102030
Change-Id: I8845b15473caca1e85306658eb7fc14289610032
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>