Long live QQmlComponent::loadFromModule!
loadFromModule is conceptually similar to loadUrl, except that it does
load the Component from a QML module URI and a typename instead of an
URL.
One big difference is that it allows loading types registered in C++,
which is not possible via URLs, as those types do not have any.

The implementation piggy-backs on the existing functionality of the
QQmlTypeLoader: We first add the URI as an import. Then we resolve the
QQmlTypeModule matching the URI.
For a C++ registered type, we can already find the existing QQmlType
there. As we don't store the QQmlTypes for composite types theree, we
might however not find the type, even if it exists. To handle that case,
we then use QQmlImports::resolveType. This is not too costly, as import
and module resolution get cached.
For composite types, we then retrieve their URL, and simply fall back to
loadUrl. This could be optimized, but is done in the simplest way for
now. For C++ registered types, we store the QQmlType in the component
instead.
Doing that allows us to use QQmlType::createWithQQmlData in
QQmlComponent::beginCreate.

Compared to loadFromUrl there are a few additional failure modes:
attempting to load singletons, querying non-existing modules, or asking
for non-existing types in a module. Those all result in component
entering an error state.

The API ignores revisioning – it is questionable whether we want to
support it, or if we can live with only resolving to the latest version.
Adding support for it at a later point would be easy; one would simply
need to pass a QTypeRevision through the various APIs.

There are two remaining tasks after this patch: We do not validate
required properties for C++ types with this approach, and we cannot load
inline components. Both will ba addressed in follow up patches.

Task-number: QTBUG-97156
Change-Id: If3ed74a499f49c668524d917089a0c237c4ddaf8
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>