QQuickMacStyle: make quickcontrols 2 react on theme change
QCocoaTheme properly detects theme/appearance changes when we switch
from 'Dark' to 'Aqua'. In widget's world this would set/propagate the
palette and all widgets will update accordingly (with styled items
having proper appearance because qmacstyle clears its cached controls).
In Quick world it's a bit different, since these are not widgets.
As a result of handling theme change, the application sends QEvent::ApplicationPaletteChange
event, which we can receive on QQuickWindow and then try to propagate
to the content of this window: set the palette on window itself (which
propagates to children) and then send event to children of QQuickRootItem,
which either ignore it or, in case of QQuickStyleItem (they don't care
about palette normally) just have to update (the cache in qquickmacstyle
is also reset on theme change notification, similar to qmacstyle).

Task-number: QTBUG-96582
Change-Id: I4191d84e15d75565fa20f609d3e7ce32a6523b2e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>