QQuickTreeViewDelegate: implement 'current' and 'selected' properties
In order for the delegate to be shown as current and/or
selected, it needs to implement the 'current' and 'selected'
properties that TableView will use.

We only change currentIndex when the user is clicking on the
the label and not the indicator. This is also how QTreeView works.
Likewise, we also only emit 'clicked' when the user is clicking on
the label, and not the indicator. Expanding and collapsing the
tree should not select the delegate at the same time.

This patch will also remove the logic (and test) that allows you to
add custom pointer handlers on top of a button, since the current
way of solving it in the delegate was not very robust. The application
should instead connect to the clicked signals coming directly from
the QAbstractButton superclass.

Change-Id: Ib07f08dfe2194303a8975082af785024da82c3e0
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>