QQmlMetaObjectPointer: optimize atomic operations [2/2]: acquires /= 2
In the destructor, as well as in the metaObject() method, the old code
used the implicit conversion from QAtomic<T> to T, which is equivalent
to load-acquire. The code performed up to two reads each, which means
two load-acquire operations where one suffices.

Cache the result of the loadAcquire() and perform further operations
on the cached value instead of the atomic variable.

Found by locally disabling QAtomic<T> -> T implicit conversion
operators.

Change-Id: I6a4ceaf462714eaad8c79fc4f26a9521525ee1c7
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>