QQmlMetaObjectPointer: optimize atomic operations [1/2]: relaxed suffice
When asserting, or in isNull() or isShared(), we don't need an acquire
fence, because the value of 'd' is all we're interested in. So just do
one relaxed atomic load, as opposed to up to two load-acquire
operations, which is what the old code used, via the implicit
conversion operator.

In the copy ctor, we can assume that external synchronization is in place
otherwise reading from other would be a data race.

Found by locally disabling QAtomic<T> -> T implicit conversion
operators.

Change-Id: I6158878a4fa9b0b511b346026f32585176b01beb
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>