doc: Add missing extension to image command
Recent changes to QDoc have made it so that QDoc does not resolve
extensions automatically for images.

That is, `cpp-qml-integration-flowchart`,
`cpp-qml-integration-flowchart.png` and
`cpp-qml-integration-flowchart.jpg` would all be treated as separate
images and being unable to be found unless they are spelled out in full.

The image command for `cpp-qml-integration-flowchart` seems to be the
only place that was affected by this change and is now changed to
explicit the extension of the image to fix the issue.

Task-number: QTBUG-100381
Change-Id: Iffabf31711c6af6a95af18f7b116bf9d18179c61
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>