CMake: Add _qmllint_json target
Adds an additional _qmllint_json target for use in CI settings and a
all_qmllint_json target to run all of them.

Since the output of ninja sub-commands can not be directly processed by
piping it into a script, the output of the qmllint command is instead
written to ${target}_qmllint.json.

[ChangeLog][qmllint][Important Behavior Changes] The --json option now has
a mandatory file argument and writes its output there. You may also pass
"-" to write it to stdout.

Fixes: QTBUG-95157
Change-Id: I3806c4113235d3aa06d3b3310e1cc7b0171b3eca
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>