Use QTestLib logging for tst_qquickfontloader_static
The test so far did not actually use QTestLib, as it has to use multiple
instances of QGuiApplication. This however confuses CI, which looks for
the log file from the test. Fix this by recording the result of the
existing test function, and then checking it in a "normal" test.

Fixes: QTBUG-102446
Change-Id: Ifdd9a2e41fe1e752e8ca0350b8706b90132675c5
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>